CDH-10665
For MapReduce indexing in *production mode* make sure to always run with the following command line flag:

-D ignoreTikaException=true

This means log exceptions that happen in the Mapper or in the Reducer, then try to continue processing, which is what production systems typically require - "just keep on trucking".

The flag is not present by default, and this is interpreted as meaning ignoreTikaException=false. This is useful for testing, where you want to fail fast when bad things happen, in order to get better diagnostics and a better debugging experience.

This is a last minute change that will still go into the private beta for search-0.1.0, as discussed in the standup today.